requirements.txt
# Machine Learning
scikit-learn>=1.0.2
numpy>=1.24.0
joblib>=1.1.0

# Geocoding
googlemaps>=4.6.0
geopy>=2.3.0

# Data Processing
pandas>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2

# Environment
python-dotenv>=0.19.0

# Utilities
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
aiofiles>=0.7.0

# Web Scraping
selenium>=4.1.0
beautifulsoup4>=4.9.3
PyPDF2>=3.0.0
webdriver_manager>=3.8.0

# HTTP Client
requests>=2.26.0
aiohttp>=3.8.0

# Progress Bars
tqdm>=4.62.0
indic-transliteration>=2.3.0

# Visualization
folium>=0.14.0
branca>=0.5.0

# Web Framework
flask>=2.0.0
